Split some integration tests to multiple files
